+++ app-api/src/main/kotlin/site/techmoa/app/article/controller/v1/ArticleController.kt
@@ -4,8 +4,8 @@ import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 import site.techmoa.app.article.controller.v1.response.GetArticleResponse
 import site.techmoa.app.common.annotation.AuthOptional
-import site.techmoa.app.common.article.ArticleContent
 import site.techmoa.app.common.article.ArticleUseCase
+import site.techmoa.app.common.article.domain.ArticleContent
 import site.techmoa.app.common.common.Page
 import site.techmoa.app.common.response.ApiResponse
