+++ app-api/src/main/kotlin/site/techmoa/app/article/controller/v1/response/GetArticleResponse.kt
@@ -1,6 +1,6 @@
 package site.techmoa.app.article.controller.v1.response
 
-import site.techmoa.app.common.article.Article
+import site.techmoa.app.common.article.domain.Article
 
 class GetArticleResponse(
     val article: Article
