+++ app-batch/app-batch-rss/src/main/kotlin/site/techmoa/app/batch/rss/RssClient.kt
@@ -1,6 +1,6 @@
 package site.techmoa.app.batch.rss
 
-import site.techmoa.app.common.article.Article
+import site.techmoa.app.common.article.domain.Article
 import site.techmoa.app.common.blog.Blog
 
 interface RssClient {
