+++ app-batch/app-batch-rss/src/main/kotlin/site/techmoa/app/batch/rss/RssReaderCollector.kt
@@ -3,8 +3,8 @@ package site.techmoa.app.batch.rss
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
-import site.techmoa.app.common.article.Article
-import site.techmoa.app.common.article.ArticlePort
+import site.techmoa.app.common.article.domain.Article
+import site.techmoa.app.common.article.port.ArticlePort
 import site.techmoa.app.common.blog.BlogPort
 import site.techmoa.app.common.blog.BlogStatus
