+++ app-batch/app-batch-rss/src/test/kotlin/site.techmoa.app.batch.rss/RssReaderCollectorTest.kt
@@ -9,7 +9,7 @@ import io.mockk.just
 import io.mockk.verify
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
-import site.techmoa.app.common.article.ArticlePort
+import site.techmoa.app.common.article.port.ArticlePort
 import site.techmoa.app.common.blog.BlogPort
 import site.techmoa.app.common.blog.BlogStatus
 import site.techmoa.app.common.comb.BlogAndArticleFixture
