+++ app-core/src/main/kotlin/site/techmoa/app/common/article/domain/ArticleContent.kt
@@ -1,4 +1,4 @@
-package site.techmoa.app.common.article
+package site.techmoa.app.common.article.domain
 
 import site.techmoa.app.common.blog.Blog
