+++ app-core/src/main/kotlin/site/techmoa/app/common/article/support/ArticleViewHandler.kt
@@ -1,7 +1,8 @@
-package site.techmoa.app.common.article
+package site.techmoa.app.common.article.support
 
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
+import site.techmoa.app.common.article.port.ArticlePort
 
 @Component
 class ArticleViewHandler(
